refactor: rename and move Logo component
