--- src/common/components/CustomIcon/AfterclassLogo.tsx
@@ -1,14 +1,0 @@
-import { CustomIcon, type CustomIconProps } from "./CustomIcon";
-
-export const AfterclassLogo = (props: CustomIconProps) => {
-  return (
-    <CustomIcon {...props}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M8.95227 2.84433C10.041 0.385225 13.959 0.385221 15.0478 2.84433L20.1616 14.3945C18.4498 13.76 16.4468 14.003 14.9184 15.203L11.9266 17.5519L11.9023 17.5324L10.0167 19.0515L9.19411 19.6973C9.07524 19.7906 8.96157 19.8879 8.85316 19.9888L7.21362 21.3096C6.65735 21.7577 6.0094 21.9732 5.36848 21.9995V22H5.27813C2.96933 22 1.38538 19.9349 2.23038 18.0263L8.95227 2.84433ZM22 19.1058C21.9927 20.6391 20.5961 22 18.7219 22H18.4488C17.8037 21.976 17.151 21.7606 16.5911 21.3096L13.171 18.5544L15.9126 16.4019C17.5204 15.1396 19.8754 15.3845 21.1726 16.9489C21.2918 17.0926 21.3978 17.2424 21.491 17.3969L21.7696 18.0263C21.8217 18.1438 21.8645 18.2619 21.8985 18.38C21.9576 18.6192 21.9913 18.8624 22 19.1058Z"
-        fill="currentColor"
-      />
-    </CustomIcon>
-  );
-};
